chore(replay): drive JSON-LD tests from shared contract
